古い探索部を削除
